apps/opportunity_finder/processing_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
kafka-python==2.0.2
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
openai==1.3.5
httpx==0.25.2